+++ core/src/main/java/com/cartoonsurvivors/game/armas/Arma.java
@@ -1,15 +1,20 @@
 package com.cartoonsurvivors.game.armas;
 
+import com.badlogic.gdx.math.Rectangle;
+import com.badlogic.gdx.math.Vector2;
+
 public abstract class Arma {
 
     private int danio;
     private float tiempoEntreAtaques;
     private float tiempoDesdeUltimoAtaque;
+    private final Rectangle areaAtaque = new Rectangle();
 
-    protected Arma(int danio, float tiempoEntreAtaques) {
+    protected Arma(int danio, float tiempoEntreAtaques,float anchoAtaque, float altoAtaque) {
         this.danio = danio;
         this.tiempoEntreAtaques = tiempoEntreAtaques;
         this.tiempoDesdeUltimoAtaque = tiempoEntreAtaques;
+        this.areaAtaque.setSize(anchoAtaque, altoAtaque);
     }
 
     public void actualizar(float delta) {
@@ -31,4 +36,22 @@ public abstract class Arma {
     public float getTiempoEntreAtaques() {
         return tiempoEntreAtaques;
     }
+
+    public Rectangle getAreaAtaque() {
+        return areaAtaque;
+    }
+
+    public void actualizarAreaAtaque(Rectangle hitboxJugador, Vector2 direccionMirada) {
+        float centroX = hitboxJugador.x + hitboxJugador.width / 2f;
+        float centroY = hitboxJugador.y + hitboxJugador.height / 2f;
+        float distancia = 50f;
+
+        float centroAtaqueX = centroX + direccionMirada.x * distancia;
+
+        float centroAtaqueY = centroY + direccionMirada.y * distancia;
+        areaAtaque.set(centroAtaqueX - areaAtaque.width / 2f, centroAtaqueY - areaAtaque.height / 2f, areaAtaque.width, areaAtaque.height);
+    }
+
+    public abstract void atacar(Rectangle hitboxJugador, Vector2 direccionMirada);
+
 }