+++ core/src/main/java/com/cartoonsurvivors/game/entidades/jugadores/Jugador.java
@@ -4,6 +4,9 @@ import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Animation;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
+import com.badlogic.gdx.math.Vector2;
+import com.cartoonsurvivors.game.armas.Arma;
+import com.cartoonsurvivors.game.armas.Espada;
 import com.cartoonsurvivors.game.entidades.Entidad;
 import com.cartoonsurvivors.game.utilidades.Constantes;
 
@@ -16,11 +19,14 @@ public abstract class Jugador extends Entidad {
     private float tiempoAnimacion = 0;
     private boolean mirandoDerecha = false;
     private boolean seEstaMoviendo = false;
+    private final Vector2 direccionMirada = new Vector2(1, 0);
+    private Arma arma;
 
     public Jugador(Texture texturaIdle, Animation<TextureRegion> animacionCaminar) {
         super(Constantes.Jugador.VIDA_INICIAL, Constantes.Jugador.VELOCIDAD_INICIAL, 0, 0, 0, 60f, 40f);
         this.texturaIdle = texturaIdle;
         this.animacionCaminar = animacionCaminar;
+        this.arma = new Espada();
     }
 
     public Texture getTexturaIdle() {
@@ -36,7 +42,12 @@ public abstract class Jugador extends Entidad {
     }
 
     public void seMueve(float direccionX, float direccionY) {
-        this.seEstaMoviendo = direccionX != 0 || direccionY != 0;
+        if (direccionX != 0 || direccionY != 0) {
+            this.seEstaMoviendo = true;
+            direccionMirada.set(direccionX, direccionY).nor();
+        } else {
+            this.seEstaMoviendo = false;
+        }
     }
 
     public void calcularLadoMirada(float direccionX) {
@@ -55,6 +66,14 @@ public abstract class Jugador extends Entidad {
 
     public boolean estaMirandoDerecha() {
         return mirandoDerecha;
+    }
+
+    public void actualizarArma(float delta) {
+        arma.actualizar(delta);
+
+        if (arma.puedeAtacar()) {
+            arma.atacar(this.hitbox, this.direccionMirada);
+        }
     }
 
     @Override